# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/users/users/verify/{key}":{"get":{"tags":["users"],"summary":"Get Verify Code","description":"Needs to be called from the email link\n:param key: The verification key from the email\n:return: Redirect to the login page","operationId":"get_verify_code_api_v1_users_users_verify__key__get","parameters":[{"required":true,"schema":{"title":"Key","type":"string"},"name":"key","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/create":{"post":{"tags":["users"],"summary":"Create User","operationId":"create_user_api_v1_users_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseUser"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseUser"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/login":{"post":{"tags":["users"],"summary":"Login","description":"Logs in a user\n:param user: The user object\n:return: The token","operationId":"login_api_v1_users_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseUser"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cli-login/token/{token}":{"get":{"tags":["cli"],"summary":"Get Token","description":"Get a token for a user.","operationId":"get_token_api_v1_cli_login_token__token__get","parameters":[{"required":true,"schema":{"title":"Token","type":"string"},"name":"token","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/list":{"get":{"tags":["projects"],"summary":"Get Projects","operationId":"get_projects_api_v1_projects_list_get","parameters":[{"required":false,"schema":{"title":"Mawoka-Auth-Header","type":"string"},"name":"mawoka-auth-header","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Projects Api V1 Projects List Get","type":"array","items":{"$ref":"#/components/schemas/BaseProject"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/create":{"post":{"tags":["projects"],"summary":"Create Project","operationId":"create_project_api_v1_projects_create_post","parameters":[{"required":false,"schema":{"title":"Mawoka-Auth-Header","type":"string"},"name":"mawoka-auth-header","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProject"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseProject"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/get/{project_name}":{"get":{"tags":["projects"],"summary":"Get Project","operationId":"get_project_api_v1_projects_get__project_name__get","parameters":[{"required":true,"schema":{"title":"Project Name","type":"string"},"name":"project_name","in":"path"},{"required":false,"schema":{"title":"Mawoka-Auth-Header","type":"string"},"name":"mawoka-auth-header","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseProject"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/update/{project_name}":{"post":{"tags":["projects"],"summary":"Update Prject","operationId":"update_prject_api_v1_projects_update__project_name__post","parameters":[{"required":true,"schema":{"title":"Project Name","type":"string"},"name":"project_name","in":"path"},{"required":false,"schema":{"title":"Mawoka-Auth-Header","type":"string"},"name":"mawoka-auth-header","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProject"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseProject"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BaseProject":{"title":"BaseProject","required":["name","data","owner","date_created","date_modified"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string","default":""},"members":{"title":"Members","type":"array","items":{"type":"string"},"default":[]},"data":{"title":"Data","type":"string"},"owner":{"title":"Owner","type":"string"},"date_created":{"title":"Date Created","type":"string"},"date_modified":{"title":"Date Modified","type":"string"}}},"BaseUser":{"title":"BaseUser","required":["password","email"],"type":"object","properties":{"password":{"title":"Password","type":"string"},"email":{"title":"Email","type":"string"}}},"CreateProject":{"title":"CreateProject","required":["name","data"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string","default":""},"members":{"title":"Members","type":"array","items":{"type":"string"},"default":[]},"data":{"title":"Data","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"UpdateProject":{"title":"UpdateProject","type":"object","properties":{"members":{"title":"Members","type":"array","items":{"type":"string"}},"description":{"title":"Description","type":"string"},"data":{"title":"Data","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}